pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "henchmen"
version = "0.1.0"
description = "AI agent factory — cloud-agnostic, villain-themed"
readme = "README.md"
license = {text = "Apache-2.0"}
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.0",
    "pydantic-settings",
    "fastapi",
    "uvicorn",
    "httpx",
    "PyGithub",
    "jinja2",
    "mcp",
]

[project.scripts]
henchmen = "henchmen.cli:main"

[project.optional-dependencies]
gcp = [
    "google-cloud-pubsub",
    "google-cloud-firestore",
    "google-cloud-storage",
    "google-cloud-run",
    "google-cloud-build",
    "google-cloud-secret-manager",
    "google-cloud-logging",
    "google-cloud-aiplatform",
    "vertexai",
    "google-genai",
    "google-adk",
]
aws = [
    "boto3>=1.34",
]
openai = [
    "openai>=1.0",
]
anthropic = [
    "anthropic>=0.30",
]
slack = [
    "slack-bolt",
]
jira = [
    "jira",
]
observability = [
    "opentelemetry-api>=1.20",
    "opentelemetry-sdk>=1.20",
    "opentelemetry-exporter-gcp-trace>=1.6",
    "opentelemetry-instrumentation-fastapi>=0.44b0",
]
all = [
    "henchmen[gcp,aws,openai,anthropic,slack,jira,observability]",
]
dev = [
    "henchmen[all]",
    "pytest",
    "pytest-asyncio",
    "pytest-cov",
    "ruff",
    "mypy",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = []

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true
